ollama_back/requirements.txt
Flask==2.2.2
torchvision==0.8.2+cu110
ollama  # 올라마를 사용하는 라이브러리